services/scraper/requirements.txt
psycopg[binary]
requests
python-dateutil